Generate README from R Markdown
I edited the README a bit, but more fundamentally, made it such that the
file can be generated from R Markdown. This has the advantage of cleanly
presenting R code snippets in the proper format and layout, as well as
help to automate future changes - such as those optimized by the 'usethis'
package.

Additonally, the introductory statement was moved to the DESCRIPTION file.